MQ2: the backend expects an int so let's cast it.

+++ include/sensors/mq2.h
@@ -11,7 +11,7 @@ public:
     MQ2Sensor(uint8_t pin, const char *board = "ESP-32", float voltageResolution = 3.3,
               uint8_t adcBitResolution = 12, const char *type = "MQ-2");
     void update();
-    float getValue();
+    int getValue();
     void begin();
     void calibrate();
